+++ src/tcp-redirections/cc-tcp-redirection.js
@@ -5,6 +5,7 @@ import { classMap } from 'lit-html/directives/class-map.js';
 import { dispatchCustomEvent } from '../lib/events.js';
 import { i18n } from '../lib/i18n.js';
 import '../atoms/cc-loader.js';
+import { defaultThemeStyles } from '../styles/default-theme.js';
 import { skeletonStyles } from '../styles/skeleton.js';
 import { waitingStyles } from '../styles/waiting.js';
 
@@ -170,6 +171,7 @@ export class CcTcpRedirection extends LitElement {
 
   static get styles () {
     return [
+      defaultThemeStyles,
       skeletonStyles,
       waitingStyles,
       // language=CSS
@@ -214,7 +216,7 @@ export class CcTcpRedirection extends LitElement {
         .text:not(.skeleton) code {
           background-color: #f3f3f3;
           border-radius: 0.25rem;
-          font-family: "SourceCodePro", "monaco", monospace;
+          font-family: var(--cc-ff-monospace);
           padding: 0.15rem 0.3rem;
         }