+++ CHANGELOG.md
@@ -6,9 +6,10 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
-...
+* Introduce `defaultThemeStyles` with monospace font-family 
 
 ## 6.2.4 (2021-04-09)
+
 * `<cc-input-number>`
   * When in controls mode and in a disabled state the buttons also appear in a disabled state 
     with a slight opacity change